Create Endpoint for googleMaps to get all Stores

+++ src/main/java/com/example/demo/models/PositionModel.java
@@ -5,10 +5,10 @@ import java.io.Serializable;
 public class PositionModel implements Serializable {
     private static final long serialVersionUID = 1L;
 
-    private Long lat;
-    private Long lng;
+    private Double lat;
+    private Double lng;
 
-    public PositionModel(Long lat, Long lng) {
+    public PositionModel(Double lat, Double lng) {
         this.lat = lat;
         this.lng = lng;
     }
@@ -20,19 +20,19 @@ public class PositionModel implements Serializable {
         return serialVersionUID;
     }
 
-    public Long getLat() {
+    public Double getLat() {
         return lat;
     }
 
-    public void setLat(Long lat) {
+    public void setLat(Double lat) {
         this.lat = lat;
     }
 
-    public Long getLng() {
+    public Double getLng() {
         return lng;
     }
 
-    public void setLng(Long lng) {
+    public void setLng(Double lng) {
         this.lng = lng;
     }
 }
